+++ eslint.config.js
@@ -1,0 +1,7 @@
+import vue3Config from '@digitalbazaar/eslint-config/vue3-recommended';
+
+export default [
+  // no longer supported
+  // 'plugin:quasar/standard',
+  ...vue3Config
+];